Move rule errors to consensus core and remove consensus error
