+++ consensus/core/src/errors/mod.rs
@@ -1,0 +1,3 @@
+pub mod block;
+pub mod coinbase;
+pub mod tx;
