+++ consensus/core/src/lib.rs
@@ -8,6 +8,8 @@ pub mod block;
 pub mod blockhash;
 pub mod blockstatus;
 pub mod coinbase;
+pub mod constants;
+pub mod errors;
 pub mod hashing;
 pub mod header;
 pub mod merkle;
