+++ consensus/tests/pipeline_tests.rs
@@ -1,11 +1,10 @@
 use consensus::{
     consensus::test_consensus::{create_temp_db, TestConsensus},
-    errors::RuleError,
     model::stores::reachability::{DbReachabilityStore, StagingReachabilityStore},
     params::MAINNET_PARAMS,
     processes::reachability::tests::{DagBlock, DagBuilder, StoreValidationExtensions},
 };
-use consensus_core::{blockhash, blockstatus::BlockStatus};
+use consensus_core::{blockhash, blockstatus::BlockStatus, errors::block::RuleError};
 use futures_util::future::join_all;
 use hashes::Hash;
 use parking_lot::RwLock;
