+++ simpa/src/simulator/miner.rs
@@ -1,11 +1,11 @@
 use super::infra::{Environment, Process, Resumption, Suspension};
 use consensus::consensus::Consensus;
-use consensus::errors::{BlockProcessResult, RuleError};
 use consensus::model::stores::virtual_state::VirtualStateStoreReader;
 use consensus::params::Params;
 use consensus_core::block::Block;
 use consensus_core::blockstatus::BlockStatus;
 use consensus_core::coinbase::MinerData;
+use consensus_core::errors::block::{BlockProcessResult, RuleError};
 use consensus_core::sign::sign;
 use consensus_core::subnets::SUBNETWORK_ID_NATIVE;
 use consensus_core::tx::{
